refactoring: 전 기능 classList Method로 refactoring 진행
