+++ hidden-header-plans.css
@@ -1,16 +1,13 @@
 .hidden-plans {
   position: fixed;
-  display: none;
   width: 100%;
   height: 0;
   top: 0;
   left: 0;
   background-color: #232f3e;
-  display: none;
   z-index: 9;
   min-width: 1920px;
   transition: all 500ms ease-out 0s;
-  overflow: hidden;
 }
 
 /* hidden-outer-contents */
@@ -403,12 +400,12 @@ sup {
 
 /* translation */
 /* hidden-contents-plan display*/
-.trans-display-on {
+.trans-block-display {
   display: block;
 }
 
-.trans-display-off {
-  display: block;
+.trans-none-display {
+  display: none;
 }
 
 .trans-show-hidden {
