+++ prime_main.html
@@ -17,7 +17,7 @@
           class="top-head-img">
       </a>
     </div>
-    <div class="hidden-plans">
+    <div class="hidden-plans trans-none-display trans-hidden">
       <div class="hidden-inner-contents">
         <div class="plan-comparison-card">
           <div class="comparison-card-header">
